Implement edge SVG logic for equality connections
In `CustomEdge.tsx` and
`ConnectionLineComponent.tsx`, add handling for
equality connections. Calculate the coordinates of
the relevant handles and their positions on the
cell to pass to `EdgeCurve.ts`.

In `EdgeCurve.ts`, vary between an S-shaped curve
and curve with tension 0.5 depending on the
position of the handles on the cells. This
prevents cases where the line is entirely straight
and where the line must pass through the cell,
leading to a better looking connection.